requirements.txt
requests
urllib3
python-dotenv==1.0.1